+++ main_debug_qt.py
@@ -5,9 +5,9 @@ Test script
 import numpy as np
 import pynapple as nap
 from PyQt6.QtWidgets import QApplication, QWidget, QHBoxLayout, QPushButton
-import pynaviz as viz
+import uviz as viz
 import sys
-from pynaviz.controller_group import ControllerGroup
+from uviz.controller_group import ControllerGroup
 
 
 tsd1 = nap.Tsd(t=np.arange(1000), d=np.cos(np.arange(1000) * 0.1))
