+++ noxfile.py
@@ -13,7 +13,7 @@ def linters(session):
 def tests(session):
     """Run the test suite."""
     # session.log("install")
-    # session.install(".[dev]")
+    session.install(".[dev]")
     tests_path = pathlib.Path(__file__).parent.resolve() / "tests"
     video_dir = tests_path / "test_video"
     video_dir.mkdir(exist_ok=True)
