+++ src/uviz/qt/__init__.py
@@ -1,0 +1,21 @@
+from .gui import scope
+from .widget_plot import (
+    IntervalSetWidget,
+    TsdFrameWidget,
+    TsdTensorWidget,
+    TsdWidget,
+    TsGroupWidget,
+    TsWidget,
+    VideoWidget,
+)
+
+__all__ = [
+    "IntervalSetWidget",
+    "TsdFrameWidget",
+    "TsdTensorWidget",
+    "TsdWidget",
+    "TsGroupWidget",
+    "TsWidget",
+    "scope",
+    "VideoWidget"
+]
