+++ src/uviz/video/__init__.py
@@ -1,0 +1,3 @@
+from .video_plot import PlotTsdTensor, PlotVideo
+
+__all__ = ["PlotVideo", "PlotTsdTensor"]
