+++ tests/test_video_handler.py
@@ -5,7 +5,8 @@ import imageio.v3 as iio
 import numpy as np
 import pytest
 
-from uviz import PlotVideo, video_handling
+from uviz import PlotVideo
+from uviz.video import video_handling
 
 
 @pytest.fixture()
